// src/app/core/api/generated/model/sessionOrganizationMembershipResponse.ts
/**
 * Precify API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrganizationScopedAccessRole } from './organizationScopedAccessRole';
import { OrganizationType } from './organizationType';


/**
 * Membership organizacional ativo do usuário autenticado.
 */
export interface SessionOrganizationMembershipResponse { 
    /**
     * Identificador da organização do vínculo.
     */
    organizationId?: string;
    /**
     * Nome da organização.
     */
    name?: string | null;
    /**
     * Código operacional opcional da organização.
     */
    code?: string | null;
    type?: OrganizationType;
    role?: OrganizationScopedAccessRole;
}
export namespace SessionOrganizationMembershipResponse {
}