// src/app/core/api/generated/model/sessionUserResponse.ts
/**
 * Precify API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Dados do usuário autenticado para bootstrap do shell do frontend.
 */
export interface SessionUserResponse { 
    /**
     * Identificador do usuário autenticado.
     */
    id?: string;
    /**
     * Nome exibido do usuário autenticado.
     */
    name?: string | null;
    /**
     * E-mail do usuário autenticado.
     */
    email?: string | null;
}